pyproject.toml
[tool.poetry]
name = "orders"
version = "0.1.0"
description = "Orders admin system"
authors = ["Michael <michael@mail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
pydantic = {version = "^2.5.3", extras = ["email"]} 
pydantic-settings = "^2.1.0"
pyjwt = "^2.10.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.9"
psycopg2-binary = "^2.9.11"
celery = "^5.3.0"
redis = "^4.5.0"
aiokafka = "*"
# confluent-kafka = "^2.3.0"
slowapi = "^0.1.9"
sqlalchemy-utils = "^0.42.1"
bcrypt = "^5.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.1"
# pytest-cov = "^4.1.0"
httpx = "^0.26.0"
mypy = "^1.8.0"
ruff = "^0.1.14"
pytest-mock = "^3.14"
fakeredis = "^2.33.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py310"

include = ["app/**/*.py", "tests/**/*.py"]
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "migrations",
    "alembic",
]

[tool.ruff.lint]
# Rules:
# E/W - pycodestyle (стиль кода)
# F - pyflakes (ошибки)
# I - isort (сортировка импортов)
# N - pep8-naming (именование)
# UP - pyupgrade (современный Python)
# B - flake8-bugbear (потенциальные баги)
# C4 - flake8-comprehensions (list/dict comprehensions)
# SIM - flake8-simplify (упрощение кода)
select = ["E", "W", "F", "I", "N", "UP", "B", "C4", "SIM"]

ignore = [
    "E501",
    "B008",
]

[tool.ruff.lint.isort]
known-first-party = ["app"]
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder"
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "passlib.*",
    "jose.*",
    "faker.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "app.services.cache"
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]